refactor: move `Endpoint` from CLI to client
