+++ bin/miden-cli/src/info.rs
@@ -33,7 +33,7 @@ async fn print_client_stats(client: &Client<impl FeltRng>) -> Result<(), String>
 }
 
 fn print_config_stats(config: &CliConfig) -> Result<(), String> {
-    println!("Node address: {}", config.rpc.endpoint.host);
+    println!("Node address: {}", config.rpc.endpoint.0.host);
     let store_len = fs::metadata(config.store.database_filepath.clone())
         .map_err(|e| e.to_string())?
         .len();
